pyproject.toml
[tool.poetry]
name = "segateway-source-mimecast"
version = "0.1.0"
description = ""
authors = ["Ryan Faircloth <ryan@dss-i.com>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.9"
backoff = "^2.2.1"
python-dotenv = "^1.0.0"
orjson = "^3.9.3"
aiofiles = "^23.2.1"
httpx = {version = "^0.24.1", extras = ["http2", "brotli"]}
furl = "^2.1.3"
Jinja2 = "^3.1.2"
python-json-logger = "^2.0.7"

[tool.poetry.dev-dependencies]
black = "^23.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py36', 'py37', 'py38']
experimental_string_processing = true

[tool.isort]
profile = 'black'
multi_line_output = 3